requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dnspython==1.16.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Login==0.4.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.24.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.1
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pymongo==3.9.0
pyOpenSSL==19.0.0
pytz==2019.2
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.3
Werkzeug==0.16.0
